Zed update through "expr: limit intermediate type creation in shapers" by nwt
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2897, authored by @nwt,
has been merged.

expr: limit intermediate type creation in shapers

A shaper computes an output type for each input type by computing types
that result from applying each of its transforms in turn.  If a shaper
has multiple transforms, this process can create intermediate types that
remain in the type context but never appear in the output.  For inputs
containing many types, creating these intermediate types can consume
substantial CPU and memory.  Rework shaper output type computation so it
does not create intermediate types.

This is part of the fix for brimdata/zed#2874.