Zed update through "expr/agg: limit intermediate type creation in Schema" by nwt
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2899, authored by @nwt,
has been merged.

expr/agg: limit intermediate type creation in Schema

Each call to Schema.Mixin creates a new record type, but for current use
cases, all but the last are intermediate types that remain in the type
context but never appear in the input.  Creating these intermediate
types can consume substantial CPU and memory.  Rework Schema so it does
not create them.

This is part of the fix for brimdata/zed#2874.

Depends on brimdata/zed#2897.