Zed update through "Clarify non-deterministic "sort" behavior in absence of an explicit field-list" by philrz
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2902, authored by @philrz,
has been merged.

Clarify non-deterministic "sort" behavior in absence of an explicit field-list

To caution users of the behaviors shown in brimdata/zed#2900, I've added text to the `sort` documentation about when the output might be non-deterministic. I adjusted some other wording while I was in there.

Closes brimdata/zed#2900