Zed update through "fix tzngio type value formatter" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2915, authored by @mccanne,
has been merged.

fix tzngio type value formatter

This commit fixes a bug in the table output that prints the new,
binary-encoded type value as a string.  For now, the fix here is
to update the tzngio printer though this should eventually go away
and the table writer should probably use package zson.