Zed update through "pkg/storage: avoid url.Parse for bare paths in ParseURI" by nwt
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2916, authored by @nwt,
has been merged.

pkg/storage: avoid url.Parse for bare paths in ParseURI

ParseURI relies on url.Parse to identify and (on non-Windows platforms)
to parse bare paths, but this produces incorrect behavior when bare
paths contain "%", "?", or "#".  Fix by avoiding url.Parse for bare
paths.