Zed update through "driver: fix deadlock after Driver method error in run" by nwt
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2929, authored by @nwt,
has been merged.

driver: fix deadlock after Driver method error in run

In driver.run, if a Driver method returns an error, a deadlock occurs
because the deferred receive from the done channel and the send to
pullerCh in the safePull goroutine both block, but the send must
complete before it can observe the canceled context and close the done
channel.

Closes brimdata/zed#2926.