Zed update through "Make "drop" test deterministic by sorting pool names returned by "zapi ls"" by philrz
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2941, authored by @philrz,
has been merged.

Make "drop" test deterministic by sorting pool names returned by "zapi ls"

When the test automation ran for my PR brimdata/zed#2940, this unrelated test failed. In running it manually I observed that repeated runs of `zapi ls` can indeed return the pool names in different order. Maybe we'd want to change that to have it be automatically sorted, but for now I'm just proposing we sort the output via Zed to so the test passes consistently.