Zed update through "change the "over" keyword to "range" in the from clause" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2943, authored by @mccanne,
has been merged.

change the "over" keyword to "range" in the from clause

This commit changes the parser to use "range" instead of "over"
to specify a range scan over a data source, which is better in line
with customary terminology.  This will affect the app's use
of the query endpoint but that branch has not yet been merged.