Zed update through "user better key names for cloud storage objects" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2944, authored by @mccanne,
has been merged.

user better key names for cloud storage objects

This commit improves the key names for cloud storage objects to
make debugging and demo easier.  It will also help people new
to the lake design more easily understand the layout.

We also got rid of the superfulous "J" component in the commit log.