Zed update through "add metadata source in "from" operator" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2954, authored by @mccanne,
has been merged.

add metadata source in "from" operator

This commit adds a meta-query capability where lake metadata
can be queried via the query endpoint.  Instead of maintaining
a set of API endpoints for every possible introspection, we
manage a namespace of meta-query sources at the lake, pool,
and (soon) branch levels.  These names are available via
the "from" command via a new bracket syntax, e.g., foo[log]
is the "log" metadata source of pool "foo".

To this end, we ported a number of the zed lake CLI commands
to use the query API and removed several methods from
lake/api/Interface.

We also included stubs and forward references anticipating the
forthcoming work to add branches.

Closes brimdata/zed#2935