Zed update through "force -z output over zng for terminal output" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2979, authored by @mccanne,
has been merged.

force -z output over zng for terminal output

This commit makes interactive Zed CLI output much more user friendly
by overriding the default zng format when sending output to a terminal.
This means that you no longer have to specify -z when running commands
interactively, which is nice for demos as you pivot between sending
output to a pipe or a file then showing the output on the terminal.

We also deleted the unused method outputflags.Flags.InitWithFormat().

Closes brimdata/zed#2974