Zed update through "Optional Put" by mattnibs
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2967, authored by @mattnibs,
has been merged.

Optional Put

Allow expressions like `x := 1` to get interpreted as a put proc.

The big change here is a change to how the parser handles a list of expressions
without a proc (e.g. `to_lower(x), to_upper(y)`). They now get parsed as
an AmbiguousExprList proc. It is then on the semantic compiler to look
at the list of expressions and determine if the list of expressions is a
Filter, Summarize or Put proc.

Closes brimdata/zed#2850