Zed update through "pass context to proc/spill.MergeSort.Spill" by nwt
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3010, authored by @nwt,
has been merged.

pass context to proc/spill.MergeSort.Spill

In proc/spill.MergeSort.Spill, sorting the records and writing them to a
file can both take multiple seconds.  For callers, this can mean a
significant delay in responding to context cancellation.  Fix this by
passing a context.Context to proc/spill.MergeSort.Spill and checking it
for cancellation.

Together with brimdata/zed#3009, this closes brimdata/zed#2983.