Zed update through "add back -p for checkout" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3023, authored by @mccanne,
has been merged.

add back -p for checkout

This commit adds back the pool option -p but only for checkout,
which checks out the main branch of the indicated pool.
This will help with the gentle slope for new users who don't need
to initially undestand or even know about branches.

We also updated various CLI help to reflect that the -p argument
is no longer relevant to them.