Zed update through "change "zed lake use" with no args to print HEAD" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3039, authored by @mccanne,
has been merged.

change "zed lake use" with no args to print HEAD

This commit changes the "use" command to print the current HEAD
if there are no command-line args, which is useful for a quick
reminder of the branch you're on.  We also reworked the logic
a bit for better readability.