Zed update through "change defaults for ZNG validation" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3050, authored by @mccanne,
has been merged.

change defaults for ZNG validation

This commit changes the default settings for ZNG validation
on the query command to be false while on the ingest commands
to be true.  We also fixed in a bug in the load service endpoint,
where validation was not being done.