Zed update through "change "zqd" to "zed" in python client" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3095, authored by @mccanne,
has been merged.

change "zqd" to "zed" in python client

This commit gets rid of the "zqd" reference in the python client.
This will break some old examples in online content but we
will have many forthcoming python examples in the near
future.  We also clarified the example in the python
README so newcomers who know python and just want to
quickly try out Zed should have an easier time.