Zed update through "fix emitter.IsTerminal check" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3115, authored by @mccanne,
has been merged.

fix emitter.IsTerminal check

When we created package storage and a scheme for wrapping stdio,
this broke the emitter.Terminal check.  The fix is to handle
the case when stdout is wrapped in a storage.NopCloser.