Use Column Names to Determine Unique Columnsets
In the past we used the name and the type to determine if a column
set was unique. Now we just use the the names. So if there is a
schema with the same fields as another schema (regardless of order)
we will display the data in a table with column headers.