Zed update through "allow "pass" as data source in a "from" trunk" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3133, authored by @mccanne,
has been merged.

allow "pass" as data source in a "from" trunk

This commit allows a pass source to be specified in a from trunk.
The code was there to handle it in compiler/kernel but ast.Pass was
missing a method to make it a valid ast.Source.