Fix Union Encoding
When encoding a union value back into zjson, we were calling
value.toString() instead of value.serialize(). The resulted in
bad zjson being sent to the store. When it was read and decoded,
it choked and caused the error.

A test has been added that replicated the error and confirmed that
it was fixed.