Nested, Quoted Zed Field Names (#1856)
Column Names are string | string[]

The dot syntax to signify a nested field no longer works now that
field names can be any string. Now, nested field names must be an
array of string containing the field name at each level of nesting.
id.orig_h becomes ["id", "orig_h"].

Represent nested fields in header using > char

Instead of the . joining the flattened fields, use a caret symbol.
This way users are not confused if the field name it self has a dot.
Hopefully people won't put carets in their field names.

Encode Fields Better

We now encode fields by encoding the field's root record and the
path to the desired field. This is a more robust solution that
what we had before.

Fix "Filter in/not in"

In order to know which item to use for the query, we now pass
the index of the value in the array or set.

Handle Null Records in getField

If a record is null, we will return a field with a null value.