Zed update through "require "dir=" prefix on mdtest-command directory" by nwt
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3153, authored by @nwt,
has been merged.

require "dir=" prefix on mdtest-command directory

Change the syntax for specifying a working directory in an
mdtest-command block info string from "mdtest-command path/to/dir" to
"mdtest-command dir=path/to/dir" in preparation for adding a new info
string option.

Part of brimdata/zed#2857.