Handle TypeAliases in Columns Code
When grabbing the columns names from the records, the code did
not properly handle when records have an alias type instead of
a record type. There can be more robust ways to test for this,
but some unit tests have been added to ensure this class of bug
doesn't pop up again.

I found another bug in the Field code where we were not handling
type aliases in determining a fields parent field.