Zed update through "disable escaping of &, <, and > in zio/jsonio.Reader" by nwt
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3177, authored by @nwt,
has been merged.

disable escaping of &, <, and > in zio/jsonio.Reader

A zio/jsonio.Reader escapes &, <, and > inside strings because it uses
encoding/json.Marshal internally.  Switch to an encoding/json.Encoder
and use its SetEscapeHTML method to disable escaping.

Closes brimdata/zed#3172.