Zed update through "fix bug where consts blocked "from" operator wiring logic" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3185, authored by @mccanne,
has been merged.

fix bug where consts blocked "from" operator wiring logic

This commit fixes a bug where type and const declarations interfered
with the logic to determine how to wire up inputs to the flowgraph.
The fix is to require all const/type declarations to appear at the
outer, top block and pull them out at the very start of compilation.

This will be reworked when we add modules, subqueries, UDFs, etc.