Zed update through "index.Finder.Nearest: support lt and gt" by mattnibs
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3196, authored by @mattnibs,
has been merged.

index.Finder.Nearest: support lt and gt

Change the name of closet to nearest, make it a public method and
support > and < operations. Get rid of ClosetGTE/LTE and have users just
call nearest with the correct operator.

Part of brimdata/zed#3183