Zed update through "reorient zbuf.Batch around []zed.Value" by nwt
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3212, authored by @nwt,
has been merged.

reorient zbuf.Batch around []zed.Value

The zbuf.Batch interface is organized around slices of pointers to
zed.Values.  Reorient it around slices of zed.Values for better cache
locality and marginally better ergonomics.  Specifically, replace the
Index, Length, and Records methods with the new Values method.

This touches many fields, parameters, and variables whose names are
variations on "record". They should all be updated to variations on
"Zed value", but to keep this diff small, I've left that for a separate PR.

For brimdata/zed#1762.

Depends on brimdata/zed#3211.