Zed update through "fix service/ztests/delete.yaml" by nwt
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3227, authored by @nwt,
has been merged.

fix service/ztests/delete.yaml

The test in service/ztests/delete/yaml is flaky because it expects
"from pool@branch:objects" to emit objects in creation order, but the
actual order is undefined.  Fix it by grabbing the first object ID
immediately after creation.