Zed update through "service request/response: Make zson the default" by mattnibs
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3242, authored by @mattnibs,
has been merged.

service request/response: Make zson the default

Change the service handlers to return zson if requested media type is
any or unset. Augment api.MediaTypeToFormat to accept a default format
that is returned if the media type header is any or unset.