Zed update through "remove TZNG and most of zio/tzngio" by nwt
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3263, authored by @nwt,
has been merged.

remove TZNG and most of zio/tzngio

TZNG has served its purpose.  Now it's just getting in the way of CSV
detection in zio/anyio.

Other packages depend on what remains of zio/tzngio.  We'll unwind them
and remove the remnants later.