change "Add Workspace..." to "Add Lake..."
This is an incomplete fix for issue #1922 but we are getting this
small change in now to immediately improve the cloud onboarding
terminology.  This will create some inconsistency in the docs that
refer to Workspace instead of Lake and throughout the code until
issue #1922 is addressed, but that should happen soon.