Zed update through "allow arbitrary top-level values in ZSON" by nwt
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3300, authored by @nwt,
has been merged.

allow arbitrary top-level values in ZSON

Relax the zson.Reader requirement that top-level values must be records.
Other parts of the system, including zio/zngio, continuee to enforce
this requirement for now.

Note that the ZSON parser does not accept consecutive top-level type
values because it interprets values after the first as decorators. See brimdata/zed#3299.