New Zealot (zed, zjson) (#1983)
In this PR, I add the @brimdata/zealot dependency, renamed the zealot directory to zealot-old, removed the zed directory and zjson file from the zealot-old directory and replaced them with the zed and zjson exported from the @brimdata/zealot package.

This is the first step in replacing zealot. Next will be to replace the client.