Zed update through "cache open pools in lake.Root" by nwt
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3318, authored by @nwt,
has been merged.

cache open pools in lake.Root

Calls to lake.OpenPool are expensive because it must open the pool's
branch and commit stores.  Add a cache of open pools in
lake.Root.poolCache to avoid reopening those stores repeatedly.

For brimdata/zed#3002.