Zed update through "add yield operator and drop "this" as an lval" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3326, authored by @mccanne,
has been merged.

add yield operator and drop "this" as an lval

This commit adds the "yield" operator which takes one or more
expressions and produces the values as "this" on the output.
Put and cut are now exclusively record operators and no longer
allow "this" as an lvalue (since yield is more natural for this).