Stricter Typescript (#1984)
Instead of fixing all 2000 errors that appear when enabling strict: true, let's do it in parts. This PR enables some of the easy strict options. The next PR will be strictNullChecks, and the last will be noImplicityAny.

Enables the following options in our typescript config:

strictFunctionTypes
strictBindCallApply
noImplicitThis
alwaysStrict
We had about 30 errors that needed to be fixed as a result of turning these one.