Zed Inspector (#1981)
The Zed Inspector is able to easily view heterogeneous data with complex and nested values.

This is a large feature that needs time to mature before we release it, so it's behind a feature flag. To run it, checkout this branch, then

npm run start -- --feature-flags=zed-inspector
The zed data is passed to the <Inspector /> component in the values prop. That component transforms the array of zed values into an array of RowData. That row data is passed into the windowing component to be virtualized. The expanded state is kept in a JavaScript Map as value => boolean.

Collateral Changes:

Some of this large diff are from these.

All the Zed value classes had to have their TypeScript types changed from a union of concrete class types to an interface type.
Moved all the Zed type colors into a single css file.
Added Tab.useState(key, defaultValue) as a way to easily put temporary state variables in the Tab section of the redux store. It can be used just like React.useState.

Known Bugs:

When the values array is large, the UI is sluggish. This is because the transformation from values into row data happens each time a value is expanded.
Expanding a type expands all instances of that type in the results. This is because there is only ever one instance of a unique zed type and the expanded state is stored in a map where the zed value is the key. Since type values all share the same identity, expanding in one place expands them everywhere.