Zed update through "Allow aggregation functions in expression context" by mattnibs
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3305, authored by @mattnibs,
has been merged.

Allow aggregation functions in expression context

Allow the use of aggregation functions in the expression context. These
so called "streaming aggregations" will output the current result of the
stateful aggregation each time a new value is added to it.

Part of brimdata/zed#3018