Zed update through "fix join when first left key value is null" by nwt
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3333, authored by @nwt,
has been merged.

fix join when first left key value is null

The join operator hangs if the first value encountered for the left join
key is null.  Fix this by checking that join.Proc.joinKey has been set
before using it.