Zed update through "Add SeekStride to pool configuration" by mattnibs
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3337, authored by @mattnibs,
has been merged.

Add SeekStride to pool configuration

Instead of having SeekStride as a global variable, move this to an
option in pools.Config. Drop -seekstride as a flag in zed command load
and add -seekstride as a hidden flag in zed create.