Zed update through "refactor zq packaging and add some fit and finish" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3331, authored by @mccanne,
has been merged.

refactor zq packaging and add some fit and finish

This commit refactors zq so we can build it in an outside repo.
While we were at it, we cleaned up the online help and changed
the -D flag to -split, and added queryflags.  We also renamed
emitter.Dir to emitter.Split and simplified the logic.