Zed update through "replace deprecated concept of "unset" with null" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3346, authored by @mccanne,
has been merged.

replace deprecated concept of "unset" with null

This commit replaces the deprecated concept of unset with Zed null.
Zed does not have unset record fields as a record with an unset
field is simply a different record type without that field.

We also removed the null check from zed.AppendInt and moved it
into zcode where it belongs.