Zed update through "change zbuf.Statser to Meter and ScannerStats to Progress" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3373, authored by @mccanne,
has been merged.

change zbuf.Statser to Meter and ScannerStats to Progress

In working on the update for package driver, I realized the
"stats" terminology for scanner status could be improved by
clarifying the design intent of a "progress meter".  So this
commit changes Statser to Meter and ScannerStats to Progress.

This is just name changes and doesn't change any logic.