Zed update through "Cors support" by mason-fish
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3374, authored by @mason-fish,
has been merged.

Cors support

Adds cors support to the api. The `OPTIONS` method must be added to any endpoints which may trigger a [pre-flight](https://developer.mozilla.org/en-US/docs/Glossary/Preflight_request) request (as opposed to a [simple](https://developer.mozilla.org/en-US/docs/Web/HTTP/CORS#simple_requests) request). For now, as we play with integrations on observablehq, I've only found it to be necessary for the /query route.

I also baked in the observable notebook's domain into the allowed origins. We may want to refactor this to be a configuration defined by cli flags instead though.