Zed update through "handle primitives in mapper to fix ZNG bug" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3392, authored by @mccanne,
has been merged.

handle primitives in mapper to fix ZNG bug

The zngio reader applies the type mapper to all top-level values
but the type mapper presumes types are complex.  This commit changes
the type mapper to check for primitives and do nothing so that
any code that uses a mapper doesn't have to special-case primitives.

Fixes brimdata/zed#3391