fast inspector (#2005)
This iteration of the inspector component removes the synchronous loop over all the zed values to create the rows. It instead just performs a count of how many rows there "will be". It does this by looking at the zed types and the "isExpanded" state. This was fast in my tests at 40ms for 10,000 values. It then looks at the scrollTop and height of the scroll container to determine which rows to actually create. To boost it further, it saves the result of the built rows in a cache when scrolling so that when it needs to re-render on a scroll, it only creates the rows that haven't been built in previous renders.

The old inspector was slow when expanding and collapsing nested values if the number of zed values was large. This is not a problem anymore.

 Fix Infinite Scrolling
 Account for type values
 Remove the "InspectorContext" class
 Fix right clicking the values
 Remove the local tab state idea
 Remove rootIndexValue and maybe the other added prop to InspectArgs
 Correctly switch to the view when the viewer search completes