Zed update through "treat Zed integer as nanoseconds in cast to duration, time" by nwt
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3448, authored by @nwt,
has been merged.

treat Zed integer as nanoseconds in cast to duration, time

When a Zed integer is cast to a duration or time, we treat the integer
value as seconds and multiply it by 10^9 to convert to nanoseconds.
Other systems represent time values as integers using sub-second units,
and this scaling makes working with those values cumbersome.  Make life
easier by treating a Zed integer value as nanoseconds in a cast to
duration or time.