Zed update through "shorten sleep time to speed up some ztests" by nwt
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3449, authored by @nwt,
has been merged.

shorten sleep time to speed up some ztests

"make test-system" runs a number of script-style ztests that poll for a
service to be ready, sleeping for 500 or 1000 milliseconds between
attempts.  Reduce that to 100 milliseconds to speed things up.  (This
shaves at least two seconds off "make test-system" for me.)