Zed update through "Rename function trunc to bucket" by mattnibs
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3450, authored by @mattnibs,
has been merged.

Rename function trunc to bucket

This change also makes it so if first argument is a duration, the
duration type is kept and is truncated by the second duration argument.

Part of brimdata/zed#2213