Zed update through "fix bug that confused ZNG error and JSON error in query end point" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3469, authored by @mccanne,
has been merged.

fix bug that confused ZNG error and JSON error in query end point

This commit fixes a bug that is tripped up by the ZNG 2022 format
changes where the query handler could return a ZNG-encoded error
instead of a JSON error but the client expected JSON.  Also, it was
hard to track down because of the vague error message, so the
error messages are improved here too.