Zed update through "refactor "this" representation in AST" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3506, authored by @mccanne,
has been merged.

refactor "this" representation in AST

This commit changes the representation of "this" to a normal
identifier in the AST and eliminates the ast.This.  We also changed
dag.Path to dag.This as this feels like a more sensible name.

This PR sets us up to experiment with binding "this" to different
scopes and binding "as" names to this in a subscope.