Zed update through "add scoping of "this" to Zed language" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3509, authored by @mccanne,
has been merged.

add scoping of "this" to Zed language

This commit introduces an "as" clause to a scoped over, which
allows "this" to be renamed so the inner scope can access this
via the name and also access the outer this via the name "this".
The scoping technique here works recursively into a hierarchy
of nested scopes.

We also fixed some problems with the semantic analysis of field
l-values and eliminated redundant code in the analysis of
the dot operator.  This was needed to correctly drop in scoping
for "this".