Zed update through "move package proc to runtime/op" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3601, authored by @mccanne,
has been merged.

move package proc to runtime/op

This commit does the first move to change proc to operator by moving zed/proc to zed/runtime/op.  There is still more work to do as outlined in brimdata/zed#3605.