Zed update through "reorganize docs" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3608, authored by @mccanne,
has been merged.

reorganize docs

This commit updates the docs directory with a simpler to understand
arrangement.  The reference docs in docs/zq/reference.md is mostly
complete while a subsequent PR will clean up the mess above.