Zed update through "release: use goreleaser" by mattnibs
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3619, authored by @mattnibs,
has been merged.

release: use goreleaser

Use goreleaser to create release assets, publish them to github as a
release and update the brew taps in github.com/brimdata/homebrew-tap to
the latest version.

This change also adds arm64 builds to releases for linux, macos, and
windows.