Remove Babel (#2211)
This PR cleans up our repo by:

Moving all source code that needs to be compiled into src
Moving the playwright tests into packages/e2e-tests
Replacing babel with swc.

swc is about 100 times faster than babel. It is designed to be a drop in replacement, and it was for the most part. I chose it for that reason as apposed to esbuild which is designed to replace a bundler like webpack. We can use esbuild later to build certain packages.

Moving all the source code into the single folder src makes all our configuration files so much simpler.