Style polish (#2225)
I improved some color and spacing styles.

I have been wondering about how to organize colors for a long time now. We use color names at the moment. However, I propose we move to descriptions of where the color will be used. The reason for this is that it's easier to understand what color you are changing, and where it is used.

This is how Slack and VSCode create custom themes, they give names to all the colors based on where they are used. I think this is what we'll do going forward. This will be easier to manage in a design tool as well.

https://code.visualstudio.com/api/references/theme-color
https://slack.com/help/articles/205166337-Change-your-Slack-theme