Inspector Update (#2243)
The bug described above was due to the way we store if a given container value is expanded or not. We used to the zed data instance itself. This worked fine most of the time. However, a feature of zed is each unique zed type is a singleton across all values that it has. This lets you compare the types of things using strict equality val1.type === val2.type. Given this feature, when you expanded one type value, all the other rows with that type also expanded.

To fix, I changed form storing the instance to storing an "indexPath" to that object. The example below shows how index paths are created.

The inspector 

Row 0 {...}
Row 1 {...}
Row 2 {name: "james", interests: ["hiking", "swimming", "coding"]}

The index path of the "interests" array is [2,1]. The first number is the row index, the second number is the field index for its record container. The benefits of this are that each value can have a unique index path, and the index paths are created on demand as you expand containers, as opposed to a solution where we'd need to walk all the nested structures.

Other improvements:

A better render preview of complex types. Before a record with an array would look like {my_array: Array(5)}, and now it looks like {my_array: ["one", "two" ...+3 ]}

Much better code structure. When I needed to add a new data type before, I had to change at least three different files each with their own switch statement that checked the type. Now each data type has it's own class that defines how it's rendered. There is now a single switch statement that checks the value and creates the view for it. For example, this is the StringView.