Zed update through "add calculator-style usage for zq" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3693, authored by @mccanne,
has been merged.

add calculator-style usage for zq

This commit changes the zq command parser to accept a single
argument that consists of a yield operator (or implied yield)
followed by optional pipelines operators so that the common
pattern of running 'echo null | zq "..." -' can be replaced simply
with 'zq "..."'.  e.g., zq '1+1' prints out 2.