Zed update through "interpret lone zq arg as query if it begins with file, from, get" by nwt
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3704, authored by @nwt,
has been merged.

interpret lone zq arg as query if it begins with file, from, get

zq fails with a "no such file" error when run with a lone argument that
is a valid query beginning with file, from, or get.  Make that work.