Update Tab Behavior (#2247)
Clicking the new tab button or cmd+t now calls the newTab() thunk. If there is not lake, go to the lakesPath, if there is but no pools, the import path, if there are some pools the new query page.

Clicking on a item in the sidebar now calls Tabs.activateByUrl finds a tab with the same location.pathname and activates it. If there is no tab, it makes a new one.

The submitSearch function does not run when you open a new draft query, or when you click on a query in the sidebar.