Zed update through "fix premature EOS in scoped over operator" by nwt
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3729, authored by @nwt,
has been merged.

fix premature EOS in scoped over operator

The flowgraph for a scoped over operator (i.e., "over ... => (...)")
shuts down prematurely if the scoped flowgraph produces no output for a
given input, as with a where operator.  Fix by adding a channel in
runtime/op/traverse.Scope called parentEOSCh, sending to it when a
Scope's parent returns an EOS, trying to receive from it when
Exit.pullPlatoon pulls an empty platoon, and interpreting the empty
platoon as EOS only if the receive succeeds.

This seems to work but doesn't feel particularly robust.  I think we
might need a different design here if we want something that's more
obviously correct.