Zed update through "add -j flag to zed and zq as a shortcut for -f json" by nwt
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3752, authored by @nwt,
has been merged.

add -j flag to zed and zq as a shortcut for -f json

Our plan was to make `-j` a shortcut for `-i json -f json`, but paring it back to just `-f json` made both the implementation (given the separation between `cli/inputflags` and `cli/outputflags`) and documentation easier. But if anyone feels strongly about it, I can add `-i json` either here or in another PR.