Zed update through "rework expression compilation to use kernel.Builder" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3795, authored by @mccanne,
has been merged.

rework expression compilation to use kernel.Builder

This commit adapts the compiler to use methods on kernel.Builder
for compiling expressions.  This will be used in a subsequent PR
to allow dataflow operators to appear inside of expressions.
We also moved CompileAssignments to package expr and renamed it
NewAssignments since it was just creating runtime objects and not
actually doing any compilation.