Zed update through "zed: delete by predicate" by mattnibs
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3791, authored by @mattnibs,
has been merged.

zed: delete by predicate

Add -predicate argument to zed delete command. The value of -predicate
is a simple zed predicate the must compare against the pool key of the
selected pool. If the flag is specified a commit will be created that
deletes all data that evaluates true for the expression.