Zed update through "simplfiy lake/api for external use" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3810, authored by @mccanne,
has been merged.

simplfiy lake/api for external use

This commit simplifies package lake/api by not exporting
the implementations and making it easier to open a lake
with either a string URL or parsed storage.URI.  We added
the OpenLake method which figures out if a local or remote
instance should be used based on the URI.