Zed update through "Remove horizontal rules from docs" by jameskerr
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3837, authored by @jameskerr,
has been merged.

Remove horizontal rules from docs

Many pages in the docs site have a horizontal rule under the heading. This breaks the automatic description tag used by the social networks. I don't think they are necessary as a visual style and removing them makes our twitter descriptions work automatically.

This is how it used to look:

<img width="627" alt="image" src="https://user-images.githubusercontent.com/3460638/165138405-b575bf96-2456-4223-b8a3-74b2997795fd.png">

We want those "---" under the title to actually be a preview of the content.

If we want to control the exact description of a page, add a 'description' key in the front-matter.

<img width="1054" alt="image" src="https://user-images.githubusercontent.com/3460638/165138569-7ed6b15f-bf39-49f9-899e-8309cfc28463.png">