Zed update through "add assert operator" by nwt
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3858, authored by @nwt,
has been merged.

add assert operator

"assert EXPR" yields the input value if the Boolean expression EXPR
evaluates to true or a structured error if it does not.  The error
structure follows that used in zed.Context.WrapError.

Closes brimdata/zed#3853.