Handle Big Arrays (#2348)
While the new inspector handles large container types with ease, the old, busted table does not. This is a quick fix so that the app doesn't crash on large arrays and sets. I expect to re-write the table view to incorporate much of the code from the inspector view, so I don't want to spend to much time propping up the old code.