Fix Infinite Scroll (#2355)
Results State
I created a Results state slice that is responsible for the main results. It looks like this:

{
    values: [] as zed.Value[],
    status: "INIT" as ResultsStatus,
    page: 1,
    perPage: 500,
    aggregationLimit: 2000,
    aggregation: false,
    key: "",
    query: "*",
    error: null as null | any,
  }

The Viewer state slice will now only store data related to the Table view, just like the Inspector only stores data related to the Inspector view.

Viewer Multi Selection

We had lots of unused code that was supposed to be for multi selection in the viewer. I removed it. I wrote that type of code again in react-arborist and it's a little better. So when we need it in the app, we'll copy it from react-arborist.

Infinite Scroll Algorithm

The infinite scroll now takes the head of perPage * pageNumber, then takes the tail of perPage. For aggregations we simply say you only get 2000 results.